[GStreamer] Add GstBufferMapped abstraction.
There are some similar abstractions I noticed while doing this called
`mapGstBuffer` and friends, which still seem to have a valid use-case:
wanting a buffer that is mapped for a longer lifetime without have to
keep track of the map infos separately. They could be subsumed by this
abstraction, but everytime they need to write to the buffer, they'd have
to remap the memory blocks.

This abstraction is more for one-short reads and writes saving the user
from remembering to unmap the buffer.